交互成功

+++ yummy-backend/src/main/java/nju/sephidator/yummybackend/dataobject/YummyFinanceDAO.java
@@ -19,5 +19,5 @@ public class YummyFinanceDAO {
 
     private Integer orderId;
 
-    private BigDecimal income;
+    private Double income;
 }
